split out Menu examples and update Menu story for easier readability
